+++ Source/RatioForge/BitTorrent/TorrentFile.cs
@@ -8,10 +8,13 @@ namespace BitTorrent
 
         private readonly long length;
 
+        private readonly string relativePath;
+
         internal TorrentFile(long len, string path) // : this()
         {
             this.fileInfo = new FileInfo(path);
             this.length = len;
+            this.relativePath = path.Replace('/', System.IO.Path.DirectorySeparatorChar);
         }
 
         internal long Length => this.length;
@@ -19,5 +22,7 @@ namespace BitTorrent
         internal string Path => this.fileInfo.FullName;
 
         internal string Name => this.fileInfo.Name;
+
+        internal string RelativePath => this.relativePath;
     }
 }